Finalizing Code
Three ipynb files represent the three methods used to generate dust_score data. The circle one draws circle every 2 hours. Then there are two ellipse drawing methods, one with inverted wind direction, and the other for control. The ellipses are drawn hourly and an error is fixed where the wind direction angle wasn't inverted properly. The dust_score_validation is the very first file that was created and this is kept in case the old process of data gathering should be accessed.

The code must still be made more legible.